pyproject.toml
[project]
name = "fastapi-sqlmodel-starter"
version = "2.0.0"
description = "Fss aims to be one of the best scaffold in PyWeb."
readme = "README.md"
keywords = [
  "fastapi",
  "scaffold",
  "sqlmodel",
  "tools",
  "web",
]
license = "MIT"
maintainers = [
     {name = "tyvekZhang", email = "tyvekzhang@gmail.com"}
]
authors = [
   {name = "tyvekZhang", email = "tyvekzhang@gmail.com"}
]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "aiomysql>=0.2",
  "aiosqlite>=0.21",
  "alembic>=1.15.2",
  "asyncpg>=0.30",
  "bcrypt>=4.3",
  "diskcache>=5.6.3",
  "fastapi[standard]>=0.115.12",
  "fastapi-async-sqlalchemy>=0.6.1",
  "fastapi-pagination>=0.13",
  "greenlet>=3.2",
  "jinja2>=3.1.6",
  "loguru>=0.7.3",
  "openpyxl>=3.1.5",
  "pandas>=2.2.3",
  "passlib>=1.7.4",
  "pyjwt>=2.10.1",
  "python-dotenv>=1.1",
  "python-jose[cryptography]>=3.4",
  "python-multipart>=0.0.20",
  "redis[hiredis]>=5.2.1",
  "slowapi>=0.1.9",
  "sqlalchemy-utils>=0.41.2",
  "sqlmodel>=0.0.24",
  "uvicorn>=0.34.2",
  "xlsxwriter>=3.2.3",
]

[tool.ruff]
include = ["src/**/*.py"]
exclude = ["doc/**/*.py"]
line-length = 120

[tool.coverage.run]
parallel = true
branch = true
source = ["src/common", "src/middleware", "src/main", "src/tests"]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[dependency-groups]
test = [
    "pre-commit>=4.2.0",
]
